refactor: refatora chamada ao service getAllCharacters
