+++ src/components/CharList.tsx
@@ -1,21 +1,12 @@
-import React, { useEffect, useState } from 'react';
-import { useSearchParams } from 'react-router-dom';
-import getCharacters from '../services/get/getCharacters';
+import React from 'react';
+import { Character } from '../interfaces/Services';
 import CharCard from './CharCard';
 
-export default function CharList() {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const page = searchParams.get('page') || '1';
-  const [characters, setCharacters] = useState<any>([]);
+interface ListProps {
+  characters: Character[];
+}
 
-  const fetchCharacters = async () => {
-    const { data } = await getCharacters(+page);
-    setCharacters(data.results);
-  };
-
-  useEffect(() => {
-    fetchCharacters();
-  }, []);
+export default function CharList({ characters }: ListProps) {
 
   return (
     <div className="charlist-container">
